requirements.txt
python-dotenv
cerebras_cloud_sdk
flask==3.0.0
flask-cors==4.0.0